tpc-server: Convert to mqtt client to publish status as birth or last will
A Last Will message can be configured on the subscribe() client but no
initial online "birth" message can be published through this mechanism,
requiring a conversion from simple subscribe registrar with callback to
a "full" mqtt client which is slightly more verbose/cumbersome to set
up.

This allows clients to know whether the script is up and running (which
won't be the case if the TPC-300 is disconnected).